LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/computed/computed.tentative-expected.txt

CONSOLE MESSAGE: TypeError: target.computedStyleMap is not a function. (In 'target.computedStyleMap()', 'target.computedStyleMap' is undefined)

Harness Error (FAIL), message = TypeError: target.computedStyleMap is not a function. (In 'target.computedStyleMap()', 'target.computedStyleMap' is undefined)
